projeto_kt_argentina/requirements_acomp_kt_argentina.txt
streamlit
pandas
plotly
psycopg2-binary
python-dotenv
numpy